StreamAnalytics/requirements.txt
# Streaming Analytics PostgreSQL Loader Requirements
# Core dependencies for loading and managing streaming analytics data
# Compatible with Python 3.8+ including Python 3.12

# PostgreSQL database adapter
psycopg2-binary>=2.9.7

# Data processing and utilities
python-dotenv>=1.0.0  # For environment variable management (.env file support)

# Python 3.12 compatibility fixes
setuptools>=68.0.0  # Required for Python 3.12+ compatibility
wheel>=0.40.0       # For building packages

# Optional: For enhanced data analysis
# pandas>=2.0.0        # For data manipulation and analysis
# numpy>=1.24.0        # For numerical operations

# Optional: Logging and monitoring
# structlog>=23.0.0    # For structured logging

# Optional: Development and testing
# pytest>=7.0.0       # For unit testing
# pytest-cov>=4.0.0   # For coverage reporting
# black>=23.0.0        # For code formatting
# flake8>=6.0.0        # For linting
